extract reusable code and move it to utils
