+++ src/layer/AbstractSentinelHubV3Layer.ts
@@ -28,8 +28,9 @@ import { ensureTimeout } from '../utils/ensureTimeout';
 
 import { Effects } from '../mapDataManipulation/const';
 import { runEffectFunctions } from '../mapDataManipulation/runEffectFunctions';
-import { CACHE_CONFIG_30MIN, CACHE_CONFIG_30MIN_MEMORY, CACHE_CONFIG_NOCACHE } from '../utils/cacheHandlers';
+import { CACHE_CONFIG_30MIN, CACHE_CONFIG_NOCACHE } from '../utils/cacheHandlers';
 import { getStatisticsProvider, StatisticsProviderType } from '../statistics/StatisticsProvider';
+import { fetchLayerParamsFromConfigurationService } from './utils';
 interface ConstructorParameters {
   instanceId?: string | null;
   layerId?: string | null;
@@ -113,42 +114,7 @@ export class AbstractSentinelHubV3Layer extends AbstractLayer {
     if (!this.dataset) {
       throw new Error('This layer does not support Processing API (unknown dataset)');
     }
-    const authToken = reqConfig && reqConfig.authToken ? reqConfig.authToken : getAuthToken();
-    if (!authToken) {
-      throw new Error('Must be authenticated to fetch layer params');
-    }
-    // Note that for SH v3 service, the endpoint for fetching the list of layers is always
-    // https://services.sentinel-hub.com/, even for creodias datasets:
-    const url = `https://services.sentinel-hub.com/configuration/v1/wms/instances/${this.instanceId}/layers`;
-    const headers = {
-      Authorization: `Bearer ${authToken}`,
-    };
-
-    // reqConfig might include the cache config from getMap, which could cache instances/${this.instanceId}/layers
-    // we do not want this as layer updates will not invalidate this cache, so we rather cache to memory
-    const reqConfigWithMemoryCache = {
-      ...reqConfig,
-      // Do not override cache if cache is disabled with `expiresIn: 0`
-      cache:
-        reqConfig && reqConfig.cache && reqConfig.cache.expiresIn === 0
-          ? reqConfig.cache
-          : CACHE_CONFIG_30MIN_MEMORY,
-    };
-    const requestConfig: AxiosRequestConfig = {
-      responseType: 'json',
-      headers: headers,
-      ...getAxiosReqParams(reqConfigWithMemoryCache, null),
-    };
-    const res = await axios.get(url, requestConfig);
-    const layersParams = res.data.map((l: any) => ({
-      layerId: l.id,
-      ...l.datasourceDefaults,
-      evalscript: l.styles[0].evalScript,
-      dataProduct: l.styles[0].dataProduct ? l.styles[0].dataProduct['@id'] : undefined,
-      legend: l.styles.find((s: any) => s.name === l.defaultStyleName)
-        ? l.styles.find((s: any) => s.name === l.defaultStyleName).legend
-        : null,
-    }));
+    const layersParams = await fetchLayerParamsFromConfigurationService(this.instanceId, reqConfig);
 
     const layerParams = layersParams.find((l: any) => l.layerId === this.layerId);
     if (!layerParams) {
